Correções de bugs e testes finais.

+++ src/FuncionarioHora.java
@@ -1,10 +1,10 @@
 
 public class FuncionarioHora extends Funcionario {
 	
-	private double horas;
+	private int horas;
 	private double salarioHora;
 	
-	public FuncionarioHora(String nome, String sobrenome, String cpf, double horas, double salarioHora) {
+	public FuncionarioHora(String nome, String sobrenome, String cpf, int horas, double salarioHora) {
 		super(nome, sobrenome, cpf);
 		
 		if(horas < 0) {
@@ -20,11 +20,11 @@ public class FuncionarioHora extends Funcionario {
 	}
 
 	
-	public double getHoras() {
+	public int getHoras() {
 		return horas;
 	}
 
-	public void setHoras(double horas) {
+	public void setHoras(int horas) {
 		
 		if(horas < 0) {
 			throw new IllegalArgumentException("Horas deve ser >= 0");
@@ -54,7 +54,7 @@ public class FuncionarioHora extends Funcionario {
 	
 	@Override
 	public String toString() {
-		return String.format("%sHoras: %.2f\nSalário Hora: %.2f\nGanhos: %.2f", super.toString(), getHoras(), getSalarioHora(), calculaGanhos());
+		return String.format("%sHoras: %d\nSalário Hora: %.2f\nGanhos: %.2f", super.toString(), getHoras(), getSalarioHora(), calculaGanhos());
 	}
